Move updatable user types to /models
